fix: addressing comment and feedbacks
