+++ src/app/(Dashboard)/layout/vertical/sidebar/MenuItems.ts
@@ -67,7 +67,7 @@ import {
 
 const Menuitems: MenuitemsType[] = [
   {
-    navlabel: true,
+    navlabel: false,
     subheader: "Home",
   },
 
@@ -75,108 +75,108 @@ const Menuitems: MenuitemsType[] = [
     id: uniqueId(),
     title: "Dashboard",
     icon: IconAperture,
-    href: "/",
-    chip: "New",
-    chipColor: "secondary",
+    href: "/dashboard",
+    // chip: "New",
+    // chipColor: "secondary",
   },
   {
     id: uniqueId(),
-    title: "Sample page",
-    icon: IconAperture,
-    href: "/sample-page",
+    title: "Users",
+    icon: IconUserPlus,
+    href: "/users",
   },
 
-  {
-    navlabel: true,
-    subheader: "Other",
-  },
-  {
-    id: uniqueId(),
-    title: "Menu Level",
-    icon: IconBoxMultiple,
-    href: "/menulevel/",
-    children: [
-      {
-        id: uniqueId(),
-        title: "Level 1",
-        icon: IconPoint,
-        href: "/l1",
-      },
-      {
-        id: uniqueId(),
-        title: "Level 1.1",
-        icon: IconPoint,
-        href: "/l1.1",
-        children: [
-          {
-            id: uniqueId(),
-            title: "Level 2",
-            icon: IconPoint,
-            href: "/l2",
-          },
-          {
-            id: uniqueId(),
-            title: "Level 2.1",
-            icon: IconPoint,
-            href: "/l2.1",
-            children: [
-              {
-                id: uniqueId(),
-                title: "Level 3",
-                icon: IconPoint,
-                href: "/l3",
-              },
-              {
-                id: uniqueId(),
-                title: "Level 3.1",
-                icon: IconPoint,
-                href: "/l3.1",
-              },
-            ],
-          },
-        ],
-      },
-    ],
-  },
-  {
-    id: uniqueId(),
-    title: "Disabled",
-    icon: IconBan,
-    href: "",
-    disabled: true,
-  },
-  {
-    id: uniqueId(),
-    title: "SubCaption",
-    subtitle: "This is the sutitle",
-    icon: IconStar,
-    href: "",
-  },
+  // {
+  //   navlabel: true,
+  //   subheader: "Other",
+  // },
+  // {
+  //   id: uniqueId(),
+  //   title: "Menu Level",
+  //   icon: IconBoxMultiple,
+  //   href: "/menulevel/",
+  //   children: [
+  //     {
+  //       id: uniqueId(),
+  //       title: "Level 1",
+  //       icon: IconPoint,
+  //       href: "/l1",
+  //     },
+  //     {
+  //       id: uniqueId(),
+  //       title: "Level 1.1",
+  //       icon: IconPoint,
+  //       href: "/l1.1",
+  //       children: [
+  //         {
+  //           id: uniqueId(),
+  //           title: "Level 2",
+  //           icon: IconPoint,
+  //           href: "/l2",
+  //         },
+  //         {
+  //           id: uniqueId(),
+  //           title: "Level 2.1",
+  //           icon: IconPoint,
+  //           href: "/l2.1",
+  //           children: [
+  //             {
+  //               id: uniqueId(),
+  //               title: "Level 3",
+  //               icon: IconPoint,
+  //               href: "/l3",
+  //             },
+  //             {
+  //               id: uniqueId(),
+  //               title: "Level 3.1",
+  //               icon: IconPoint,
+  //               href: "/l3.1",
+  //             },
+  //           ],
+  //         },
+  //       ],
+  //     },
+  //   ],
+  // },
+  // {
+  //   id: uniqueId(),
+  //   title: "Disabled",
+  //   icon: IconBan,
+  //   href: "",
+  //   disabled: true,
+  // },
+  // {
+  //   id: uniqueId(),
+  //   title: "SubCaption",
+  //   subtitle: "This is the sutitle",
+  //   icon: IconStar,
+  //   href: "",
+  // },
 
-  {
-    id: uniqueId(),
-    title: "Chip",
-    icon: IconAward,
-    href: "",
-    chip: "9",
-    chipColor: "primary",
-  },
-  {
-    id: uniqueId(),
-    title: "Outlined",
-    icon: IconMoodSmile,
-    href: "",
-    chip: "outline",
-    variant: "outlined",
-    chipColor: "primary",
-  },
-  {
-    id: uniqueId(),
-    title: "External Link",
-    external: true,
-    icon: IconStar,
-    href: "https://google.com",
-  },
+  // {
+  //   id: uniqueId(),
+  //   title: "Chip",
+  //   icon: IconAward,
+  //   href: "",
+  //   chip: "9",
+  //   chipColor: "primary",
+  // },
+  // {
+  //   id: uniqueId(),
+  //   title: "Outlined",
+  //   icon: IconMoodSmile,
+  //   href: "",
+  //   chip: "outline",
+  //   variant: "outlined",
+  //   chipColor: "primary",
+  // },
+  // {
+  //   id: uniqueId(),
+  //   title: "External Link",
+  //   external: true,
+  //   icon: IconStar,
+  //   href: "https://google.com",
+  // },
 ];
 
 export default Menuitems;
